% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adjusted_rand_index}
\alias{adjusted_rand_index}
\title{Adjusted Rand Index}
\usage{
adjusted_rand_index(x, y = NULL)
}
\arguments{
\item{x}{an integer vector where x[i] representing the}

\item{y}{an integer vector}
}
\value{
A single real value.

adjusted_rand_index(x, y), adjusted_rand_index(table(x, y))
}
\description{
The Adjusted Rand Index is a pairwise similarity measure used....
Two identical partitions yield the score of 1.0.
The index is corrected for chance, i.e., its expected value is 0.0
for two independent partitions.
Due to the adjustment. the resulting index might also be negative for some
inputs.
}
\references{
Hubert L., Arabie P., Comparing Partitions,
Journal of Classification 2(1), 1985, pp. 193-218, esp. Eqs. (2) and (4)
}
