% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bonferroni}
\alias{bonferroni}
\title{The Normalised Bonferroni Index}
\usage{
bonferroni(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
The value of the inequity index, a number in [0,1].
}
\description{
The normalised Bonferroni index is given by:
\deqn{
    B(x_1,\dots,x_n) = \frac{
    \sum_{i=1}^{n}  \left( n-\sum_{j=1}^i \frac{n}{n-j+1} \right)
         x_{\sigma(n-i+1)}
    }{
    (n-1) \sum_{i=1}^n x_i
}   },

Time complexity: \eqn{O(n)} for sorted (increasingly) data.
Otherwise, the vector will be sorted.
}
\examples{
bonferroni(c(2, 2, 2, 2, 2))  # no inequality
bonferroni(c(0, 0, 10, 0, 0)) # one has it all
bonferroni(c(7, 0, 3, 0, 0))  # give to the poor, take away from the rich
bonferroni(c(6, 0, 3, 1, 0))  # (a.k.a. Pigou-Dalton principle)

}
\seealso{
gini
}
