% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cluster_validity_measures}
\alias{cluster_validity_measures}
\alias{calinski_harabasz_index}
\alias{dunnowa_index}
\alias{generalised_dunn_index}
\alias{negated_ball_hall_index}
\alias{negated_davies_bouldin_index}
\alias{silhouette_index}
\alias{silhouette_w_index}
\alias{wcnn_index}
\alias{wcss_index}
\title{Internal Cluster Validity Measures}
\usage{
calinski_harabasz_index(X, y, K)

dunnowa_index(X, y, K, M = 10L, owa_numerator = "Min", owa_denominator = "Max")

generalised_dunn_index(X, y, K, lowercase_delta, uppercase_delta)

negated_ball_hall_index(X, y, K)

negated_davies_bouldin_index(X, y, K)

silhouette_index(X, y, K)

silhouette_w_index(X, y, K)

wcnn_index(X, y, K, M = 10L)

wcss_index(X, y, K)
}
\arguments{
\item{X}{numeric matrix with \code{n} rows and \code{d} columns,
representing \code{n} points in a \code{d}-dimensional space}

\item{y}{vector of \code{n} integer labels with elements in \eqn{[1, K]},
representing a partition whose \emph{quality} is to be
assessed; \code{y[i]} is the cluster ID of the \code{i}-th point,
\code{X[i, ]}}

\item{K}{number of clusters, equal to \code{max(y)}}

\item{M}{number of nearest neighbours}

\item{owa_numerator, owa_denominator}{single string defining
the OWA operator to use in the definition of the DuNN index;
one of: \code{"Mean"}, \code{"Min"}, \code{"Max"}, \code{"Const"},
\code{"SMin:M"}, \code{"SMax:M"}, where \code{M} is an integer
defining the number of nearest neighbours.}

\item{lowercase_delta}{an integer between 1 and 6, denoting
\eqn{d_1}, ..., \eqn{d_6} in the definition
of the generalised Dunn index (numerator)}

\item{uppercase_delta}{an integer between 1 and 3, denoting
\eqn{D_1}, ..., \eqn{D_3} in the definition
of the generalised Dunn index (denominator)}
}
\value{
A single numeric value.
}
\description{
Implements a number of cluster validity indices critically
reviewed in (Gagolewski, Bartoszuk, Cena, 2021). See Section 2
therein for the respective definitions.

The greater the index value, the more \emph{valid} (whatever that means)
the assessed partition. For consistency, the Ball-Hall and
Davies-Bouldin indexes take negative values.
}
\examples{
X <- as.matrix(iris[,1:4])
X[,] <- jitter(X)  # otherwise we get a non-unique solution
y <- as.integer(iris[[5]])
calinski_harabasz_index(X, y, max(y))
calinski_harabasz_index(X, sample(1:3, nrow(X), replace=TRUE), max(y))

}
\references{
G.H. Ball, D.J. Hall,
ISODATA: A novel method of data analysis and pattern classification,
Technical report No. AD699616, Stanford Research Institute, 1965.

J. Bezdek, N. Pal, Some new indexes of cluster validity,
IEEE Transactions on Systems, Man, and Cybernetics, Part B (Cybernetics) 28
1998, pp. 301--315, \doi{10.1109/3477.678624}.

T. Calinski, J. Harabasz. A dendrite method for cluster analysis,
Communications in Statistics, 3(1), 1974, pp. 1-27,
\doi{10.1080/03610927408827101}.

D.L. Davies, D.W. Bouldin,
A Cluster Separation Measure,
IEEE Transactions on Pattern Analysis and Machine Intelligence. PAMI-1 (2),
1979, pp. 224-227, \doi{10.1109/TPAMI.1979.4766909}.

J.C. Dunn, A Fuzzy Relative of the ISODATA Process and Its Use in Detecting
Compact Well-Separated Clusters, Journal of Cybernetics 3(3), 1973,
pp. 32-57, \doi{10.1080/01969727308546046}.

M. Gagolewski, M. Bartoszuk, A. Cena,
Are cluster validity measures (in)valid?, Information Sciences 581,
620–636, 2021, \doi{10.1016/j.ins.2021.10.004};
preprint: \url{https://raw.githubusercontent.com/gagolews/bibliography/master/preprints/2021cvi.pdf}.

P.J. Rousseeuw, Silhouettes: A Graphical Aid to the Interpretation and
Validation of Cluster Analysis, Computational and Applied Mathematics 20,
1987, pp. 53-65, \doi{10.1016/0377-0427(87)90125-7}.
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski} and other contributors
}
\seealso{
The official online manual of \pkg{genieclust} at \url{https://genieclust.gagolewski.com/}

Gagolewski M., \pkg{genieclust}: Fast and robust hierarchical clustering, \emph{SoftwareX} 15:100722, 2021, \doi{10.1016/j.softx.2021.100722}.

}
