% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{emst_mlpack}
\alias{emst_mlpack}
\title{Euclidean Minimum Spanning Tree [DEPRECATED]}
\usage{
emst_mlpack(d, leaf_size = 1, verbose = FALSE)
}
\arguments{
\item{d}{a numeric matrix (or an object coercible to one,
e.g., a data frame with numeric-like columns)}

\item{leaf_size}{size of leaves in the K-d tree,
controls the trade-off between speed and memory consumption}

\item{verbose}{logical; whether to print diagnostic messages}
}
\value{
An object of class \code{mst}, see \code{\link{mst}()} for details.
}
\description{
This function is deprecated. Use \code{\link{mst}()} instead.
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski} and other contributors
}
\seealso{
The official online manual of \pkg{genieclust} at \url{https://genieclust.gagolewski.com/}

Gagolewski M., \pkg{genieclust}: Fast and robust hierarchical clustering, \emph{SoftwareX} 15:100722, 2021, \doi{10.1016/j.softx.2021.100722}.

}
