% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gclust.R
\name{gclust}
\alias{gclust}
\alias{gclust.default}
\alias{gclust.dist}
\title{The Genie++ Hierarchical Clustering Algorithm}
\usage{
gclust(d, ...)

\method{gclust}{default}(
  d,
  gini_threshold = 0.3,
  M = 1L,
  postprocess = "boundary",
  distance = "euclidean",
  ...
)

\method{gclust}{dist}(d, gini_threshold = 0.3, M = 1L, postprocess = "boundary", ...)
}
\arguments{
\item{d}{TODO}

\item{...}{TODO}

\item{gini_threshold}{TODO}

\item{M}{TODO}

\item{postprocess}{TODO}

\item{distance}{TODO}
}
\value{
A list of class \code{hclust}, see \code{\link[stats]{hclust}}.
}
\description{
TODO
}
\examples{
library("datasets")
data("iris")
X <- iris[1:4]
#h <- gclust(X)
# y_pred <- cutree(h, 3)
y_test <- iris[,5]
#plot(iris[,2], iris[,3], col=y_pred,
#   pch=as.integer(iris[,5]), asp=1, las=1)
#adjusted_rand_score(y_test, y_pred)

}
\references{
Gagolewski M., Bartoszuk M., Cena A.,
Genie: A new, fast, and outlier-resistant hierarchical clustering algorithm,
\emph{Information Sciences} 363, 2016, pp. 8-23.
}
