% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genieclust-package.R
\docType{package}
\name{genieclust-package}
\alias{genieclust-package}
\alias{genieclust}
\title{The Genie++ Hierarchical Clustering Algorithm (with Extras)}
\description{
See \code{\link{genie}()} for more details.
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski} and other contributors
}
\seealso{
The official online manual of \pkg{genieclust} at \url{https://genieclust.gagolewski.com/}
}
