% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gini}
\alias{gini}
\title{The Normalised Gini Index}
\usage{
gini(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
The value of the inequity index, a number in [0,1].
}
\description{
The normalised  Gini index is given by:
\deqn{
    G(x_1,\dots,x_n) = \frac{
    \sum_{i=1}^{n-1} \sum_{j=i+1}^n |x_i-x_j|
    }{
    (n-1) \sum_{i=1}^n x_i
    }.
}

Time complexity: \eqn{O(n)} for sorted (increasingly) data.
Otherwise, the vector will be sorted.
Note that for sorted data, it holds:
\deqn{
    G(x_1,\dots,x_n) = \frac{
    \sum_{i=1}^{n} (n-2i+1) x_{\sigma(n-i+1)}
    }{
    (n-1) \sum_{i=1}^n x_i
    },
}
where \eqn{\sigma} is an ordering permutation of \eqn{(x_1,\dots,x_n)}.
}
\examples{
gini(c(2, 2, 2, 2, 2))  # no inequality
gini(c(0, 0, 10, 0, 0)) # one has it all
gini(c(7, 0, 3, 0, 0))  # give to the poor, take away from the rich
gini(c(6, 0, 3, 1, 0))  # (a.k.a. Pigou-Dalton principle)

}
\seealso{
bonferroni
}
