% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{omp_set_num_threads}
\alias{omp_set_num_threads}
\alias{omp_get_max_threads}
\title{Get or Set the Number of Threads}
\usage{
omp_set_num_threads(n_threads)

omp_get_max_threads()
}
\arguments{
\item{n_threads}{maximum number of threads to use;}
}
\value{
\code{omp_get_max_threads} returns the maximal number
of threads that will be used during the next call to a parallelised
function, not the maximal number of threads possibly available.
It there is no built-in support for OpenMP, 1 is always returned.

For \code{omp_set_num_threads}, the previous value of \code{max_threads}
is output.
}
\description{
These functions get or set the maximal number of OpenMP threads that
can be used by \code{\link{knn_euclid}} and \code{\link{mst_euclid}},
amongst others.
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski} and other contributors
}
\seealso{
The official online manual of \pkg{genieclust} at \url{https://genieclust.gagolewski.com/}

Gagolewski M., \pkg{genieclust}: Fast and robust hierarchical clustering, \emph{SoftwareX} 15:100722, 2021, \doi{10.1016/j.softx.2021.100722}.

}
