% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{emst_mlpack}
\alias{emst_mlpack}
\title{Euclidean Minimum Spanning Tree}
\usage{
emst_mlpack(X, verbose = FALSE)
}
\arguments{
\item{X}{a numeric matrix (or an object coercible to one,
e.g., a data frame with numeric-like columns)}

\item{verbose}{logical; whether to print diagnostic messages}
}
\value{
An object of class \code{mst}, see \code{\link{mst}()} for details.
}
\description{
Provides access to an implementation of the Dual-Tree Borůvka
algorithm based on kd-trees from MLPACK. It is fast for (very) low-dimensional
Euclidean spaces. For higher dimensional spaces (say, over 5 features)
or other metrics, use the parallelised Prim-like algorithm implemented
in \code{\link{mst}()}.
}
\references{
March W.B., Ram P., Gray A.G.,
Fast Euclidean Minimum Spanning Tree: Algorithm, Analysis, and Applications,
Proc. ACM SIGKDD'10 (2010) 603-611, \url{https://mlpack.org/papers/emst.pdf}.

Curtin R.R., Edel M., Lozhnikov M., Mentekidis Y., Ghaisas S., Zhang S.,
mlpack 3: A fast, flexible machine learning library,
Journal of Open Source Software 3(26), 726, 2018.
}
