% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gclust.R
\name{gclust}
\alias{gclust}
\alias{gclust.default}
\alias{gclust.dist}
\alias{gclust.mst}
\alias{genie}
\alias{genie.default}
\alias{genie.dist}
\alias{genie.mst}
\title{Hierarchical Clustering Algorithm Genie}
\usage{
gclust(d, ...)

\method{gclust}{default}(
  d,
  gini_threshold = 0.3,
  M = 0L,
  distance = c("euclidean", "l2", "manhattan", "cityblock", "l1", "cosine"),
  verbose = FALSE,
  ...
)

\method{gclust}{dist}(d, gini_threshold = 0.3, M = 0L, verbose = FALSE, ...)

\method{gclust}{mst}(d, gini_threshold = 0.3, verbose = FALSE, ...)

genie(d, ...)

\method{genie}{default}(
  d,
  k,
  gini_threshold = 0.3,
  M = 0L,
  distance = c("euclidean", "l2", "manhattan", "cityblock", "l1", "cosine"),
  verbose = FALSE,
  ...
)

\method{genie}{dist}(d, k, gini_threshold = 0.3, M = 0L, verbose = FALSE, ...)

\method{genie}{mst}(d, k, gini_threshold = 0.3, verbose = FALSE, ...)
}
\arguments{
\item{d}{a numeric matrix (or an object coercible to one,
e.g., a data frame with numeric-like columns) or an
object of class \code{dist} (see \code{\link[stats]{dist}}),
or an object of class \code{mst} (see \code{\link[deadwood]{mst}})}

\item{...}{further arguments passed to \code{\link[deadwood]{mst}()}}

\item{gini_threshold}{threshold for the Genie correction, i.e.,
the Gini index of the cluster size distribution;
threshold of 1.0 leads to the single linkage algorithm;
low thresholds highly penalise the formation of small clusters}

\item{M}{smoothing factor; \eqn{M \leq 1} gives the selected \code{distance};
otherwise, the mutual reachability distance is used}

\item{distance}{metric used to compute the linkage, one of:
\code{"euclidean"} (synonym: \code{"l2"}),
\code{"manhattan"} (a.k.a. \code{"l1"} and \code{"cityblock"}),
\code{"cosine"}}

\item{verbose}{logical; whether to print diagnostic messages
and progress information}

\item{k}{the desired number of clusters to detect}
}
\value{
\code{gclust()} computes the entire clustering hierarchy; it
returns a list of class \code{hclust}; see \code{\link[stats]{hclust}}.
Use \code{\link[stats]{cutree}} to obtain an arbitrary \code{k}-partition.

\code{genie()} returns a \code{k}-partition, i.e., a vector whose
\eqn{i}-th element denotes the \eqn{i}-th input point's cluster label
between 1 and \code{k}.
}
\description{
A reimplementation of \emph{Genie} - a robust and outlier resistant
clustering algorithm by Gagolewski, Bartoszuk, and Cena (2016).
The Genie algorithm is based on the minimum spanning tree (MST) of the
pairwise distance graph of a given point set.
Just like the Single Linkage method, it consumes the edges
of the MST in an increasing order of weights. However, it prevents
the formation of clusters of highly imbalanced sizes; once the Gini index
(see \code{\link{gini_index}()}) of the cluster size distribution
raises above \code{gini_threshold}, merging a point group
of the smallest size is enforced.

A clustering can also be computed with respect to the
\eqn{M}-mutual reachability distance (based, e.g., on the Euclidean metric),
which is used in the definition of the HDBSCAN* algorithm
(see \code{\link[deadwood]{mst}()} for the definition).
For the smoothing factor \eqn{M>0}, outliers are pulled away from
their neighbours.  This way, the Genie algorithm gives an alternative
to the HDBSCAN* algorithm (Campello et al., 2013) that is able to detect
a predefined number of clusters and indicate outliers (Gagolewski, 2026)
without depending on DBSCAN*'s \code{eps} or HDBSCAN*'s
\code{min_cluster_size} parameters.  Also make sure to check out
the Lumbermark method that is also based on MSTs.
}
\details{
As in the case of all the distance-based methods (including
k-nearest neighbours and DBSCAN),  the standardisation of the input features
is definitely worth giving a try.  Oftentimes, applying feature selection
and engineering techniques (e.g., dimensionality reduction) might lead
to more meaningful results.

If \code{d} is a numeric matrix or an object of class \code{dist},
\code{\link[deadwood]{mst}()} will be called to compute an MST, which generally
takes at most \eqn{O(n^2)} time. However, by default, a faster algorithm
based on K-d trees is selected automatically for low-dimensional Euclidean
spaces; see \code{\link[quitefastmst]{mst_euclid}} from
the \pkg{quitefastmst} package.

Once a minimum spanning tree is determined, the Genie algorithm runs in
\eqn{O(n \sqrt{n})} time.  If you want to test different
\code{gini_threshold}s or \code{k}s, it is best to compute
the MST explicitly beforehand.

Due to Genie's original definition, the resulting partition tree (dendrogram)
might violate the ultrametricity property (merges might occur at levels that
are not increasing w.r.t. a between-cluster distance).
\code{gclust()} automatically corrects departures from
ultrametricity by applying \code{height = rev(cummin(rev(height)))}.
}
\examples{
library("datasets")
data("iris")
X <- jitter(as.matrix(iris[2:3]))
h <- gclust(X)
y_pred <- cutree(h, 3)
y_test <- as.integer(iris[,5])
plot(X, col=y_pred, pch=y_test, asp=1, las=1)
adjusted_rand_score(y_test, y_pred)
normalized_clustering_accuracy(y_test, y_pred)

y_pred2 <- genie(X, 3, M=5)  # the 5-mutual reachability distance
plot(X[,1], X[,2], col=y_pred2, pch=y_test, asp=1, las=1)
is_outlier <- is.na(y_pred2)
points(X[is_outlier, ], col="gray", pch=10)
normalized_clustering_accuracy(y_test[!is_outlier], y_pred2[!is_outlier])

}
\references{
Gagolewski M., Bartoszuk M., Cena A.,
Genie: A new, fast, and outlier-resistant hierarchical clustering algorithm,
\emph{Information Sciences} 363, 2016, 8-23,
\doi{10.1016/j.ins.2016.05.003}

Campello R.J.G.B., Moulavi D., Sander J.,
Density-based clustering based on hierarchical density estimates,
\emph{Lecture Notes in Computer Science} 7819, 2013, 160-172,
\doi{10.1007/978-3-642-37456-2_14}

Gagolewski M., Cena A., Bartoszuk M., Brzozowski L.,
Clustering with minimum spanning trees: How good can it be?,
\emph{Journal of Classification} 42, 2025, 90-112,
\doi{10.1007/s00357-024-09483-1}

Gagolewski M., genieclust: Fast and robust hierarchical clustering,
\emph{SoftwareX} 15, 2021, 100722, \doi{10.1016/j.softx.2021.100722}

Gagolewski M., deadwood, in preparation, 2026

Gagolewski M., quitefastmst, in preparation, 2026
}
\seealso{
The official online manual of \pkg{genieclust} at \url{https://genieclust.gagolewski.com/}

Gagolewski, M., \pkg{genieclust}: Fast and robust hierarchical clustering, \emph{SoftwareX} 15:100722, 2021, \doi{10.1016/j.softx.2021.100722}

\code{\link[deadwood]{mst}()} for the minimum spanning tree routines

\code{\link{normalized_clustering_accuracy}()} (amongst others) for external
cluster validity measures
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski} and other contributors
}
