% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genieclust-package.R
\docType{package}
\name{genieclust-package}
\alias{genieclust-package}
\alias{genieclust}
\title{The Genie++ Hierarchical Clustering Algorithm (with Extras)}
\description{
See \code{\link{genie}()} for more details.
}
\author{
Marek Gagolewski
}
